requirements.txt
# CBT (Core Brim Tech) - Comprehensive Requirements
# Python 3.6+ required

# Core Dependencies
pycurl>=7.45.0
requests>=2.31.0
urllib3>=2.0.0

# Cryptography & Security
pycryptodomex>=3.19.0
cryptography>=41.0.0

# Network & Protocol Libraries
scapy>=2.5.0
dnspython>=2.4.0

# GeoIP & Location Services
GeoIP>=1.3.2
python-geoip>=1.2
pygeoip>=0.3.2

# Web & HTTP
h2>=4.1.0  # For HTTP/2 attacks (optional but recommended)
httpx>=0.25.0  # Modern HTTP client (optional)

# Search & OSINT
duckduckgo-search>=3.9.0
whois>=0.9.0

# Data Processing
python-whois>=0.8.0

# AI & Machine Learning (for AI features)
numpy>=1.24.0  # For AI calculations
scikit-learn>=1.3.0  # For ML-based evasion

# Network Scanning & Discovery
python-nmap>=0.7.1  # For port scanning
shodan>=1.30.0  # Optional: Shodan API integration
censys>=2.2.0  # Optional: Censys API integration

# Async & Performance
aiohttp>=3.9.0  # For async HTTP requests
asyncio>=3.4.3  # Built-in but ensure compatibility

# Data Storage & Serialization
pymongo>=4.6.0  # Optional: MongoDB for botnet data storage
redis>=5.0.0  # Optional: Redis for caching

# Logging & Monitoring
colorama>=0.4.6  # Colored terminal output
tqdm>=4.66.0  # Progress bars

# Web Interface Dependencies (if using enhanced GUI)
flask>=3.0.0  # Optional: Enhanced web interface
flask-socketio>=5.3.0  # Optional: Real-time updates

# System Dependencies (install via apt/yum)
# libpython3.11-dev
# python3-pycurl
# python3-geoip
# python3-whois
# python3-cryptography
# python3-requests
# libgeoip1
# libgeoip-dev

# Development Dependencies (optional)
# pytest>=7.4.0
# black>=23.11.0
# flake8>=6.1.0
